refactor: Add linksHelpers
